### 2023-04-20 #5390 @editorialbot
[PRE REVIEW]: APECSS: A software library for cavitation bubble dynamics and acoustic emissions
**Submitting author:** @fabiandenner (<a href="http://orcid.org/0000-0001-5812-061X">Fabian Denner</a>)
**Repository:** https://github.com/polycfd/apecss
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.3
