### 2023-04-20 #5390 @fabiandenner
The two references flagged with a missing DOI don't, as far as I was able to determine, have a DOI.
