### 2023-05-17 #5435 @fabiandenner
@svchb I agree, a direct comparison of APECSS to similar software tools would be a nice addition. On multiple occasions, also just before submitting to JOSS as we anticipated this question, we tried our best to identify open-source tools with similar capabilities. However, we also couldn't find any. We validated APECSS thoroughly, especially in our [Physics of Fluids paper](https://doi.org/10.1063/5.0131930) published earlier this year, and this can also be checked with the examples provided in the repository (we mention the literature case it should reproduce in the README.md accompanying each example). In terms of performance, we could conduct a comparison to an in-house Python implementation, using SciPy functions (with the same Runge-Kutta method) to solve the governing ODE, for a few examples. I'm not sure how representative such an implementation would be, but it could at least give a rough orientation about the performance benefit. What do you think?
