### 2023-05-17 #5435 @svchb
@fabiandenner I had not yet time to check your code or examples. Anyway I am not sure how representative that would be as performance of Python based solvers largely depends on avoiding native Python code execution of computational intensive parts.
