### 2023-05-17 #5435 @fabiandenner
@svchb I agree, our Python scripts are perhaps best described as a simple implementation with off-the-shelf components, not a sophisticated or performance-optimized implementation.
