### 2023-05-17 #5435 @fabiandenner
@svchb Yes it is supposed to perform time steps. The [README.md](https://github.com/fabiandenner/apecss/blob/main/examples/gastemperature/README.md) of this test case provides an example of the execution command (that corresponds to reproducing the references literature result). I think what's missing in your execution are the excitation frequency and amplitude, and the end time of the simulation.

I'll expand the Quick Start section of the main README.md file to include the additional steps to running one of the examples.

The help arguments is a nice idea, thanks for this suggestion. I'll add this. Also, I'll have a look into the error messages.
