### 2023-05-17 #5435 @fabiandenner
@svchb I just committed some small changes. I expanded the Quick Start Guide in the primary README.md and moved it further up in the text, so that it is better visible. Also, -h now provides some basic command line help.
