### 2023-05-24 #5435 @fabiandenner
@svchb and @jmansour Just to let you know, I found a small mistake in the paper draft. The label of the $y$-axis in Figure 1(h) was incorrect, $u(r,t)$ instead of $u(r)$. I committed a corrected a figure.
