### 2023-05-31 #5435 @svchb
Some comments:
- Your in-code documentation is lacking it would be better to document major functions within the code
- It would be nice if you add scripts to plot the output of your examples
- Your documentation file would benefit from more diagrams/illustrations/plots

Tests:
- Your tests only verify that the code compiles and runs. Please adjust the tests in a way that they compare to a verified result as well.

Besides this testing aspect I am satisfied with the paper, documentation and code.
