### 2023-05-31 #5435 @fabiandenner
@svchb Thanks for the feedback. Following your suggestions I'll add:
- in-code documentation for major functions and source files.
- scripts to plot the results, together with reference results where applicable. Many of the examples reproduce results found in the literature (indicated in the corresponding README files), which can be used to validate the results. However, I agree with you, this might be a little cumbersome at the moment.
- sample results (produced with the scripts mentioned before) and illustrations to the pdf-documentation.

I'll try to find the time to add this by the end of the week.

Thank you very much for having taken the time to review our software and for the helpful feedback, we greatly appreciate it.
