### 2023-06-02 eac6fdf
Scripts to plot the results of the examples (incl. reference results where applicable)
