+++ examples/ultrasound/plot_result_lipidcoated_emissions.py
@@ -1,0 +1,75 @@
+import numpy as np
+import matplotlib.pyplot as plt
+
+plt.rcParams['font.family']='serif'
+plt.rcParams['font.serif']=['Times New Roman'] + plt.rcParams['font.serif']
+plt.rcParams['mathtext.fontset']='stix'
+plt.rcParams['font.size']=10
+
+cm = 1/2.54
+
+Bubble = np.genfromtxt("Gilmore_R1.000e-06_fa1.000e+06_pa6.000e+05.txt", delimiter=" ")
+pLmax = np.genfromtxt("EmissionsNode_pLmax.txt", delimiter=" ")
+
+fig1 = plt.figure(figsize=(17*cm,10*cm))
+ax1 = plt.subplot2grid((2,9),(0,0),colspan=3)
+ax2 = plt.subplot2grid((2,9),(0,3),colspan=3)
+ax3 = plt.subplot2grid((2,9),(0,6),colspan=3)
+ax4 = plt.subplot2grid((2,9),(1,0),colspan=3)
+ax5 = plt.subplot2grid((2,9),(1,3),colspan=3)
+plt.subplots_adjust(wspace=30*cm,hspace=1.2*cm)
+
+ax1.set(xlabel=r'$t$ [$\mu$s]',ylabel=r'$R(t)$ [$\mu$m]')
+ax1.set_xlim(xmin=6,xmax=11)
+ax1.grid(color='gainsboro', linestyle='-', linewidth=0.5)
+ax1.plot(Bubble[:, 1]*1e6, Bubble[:, 3]*1e6, linestyle='solid', linewidth=0.75,color='navy')
+
+ax2.set(xlabel=r'$t$ [$\mu$s]',ylabel=r'$\dot{R}(t)$ [m/s]')
+ax2.set_xlim(xmin=6,xmax=11)
+ax2.grid(color='gainsboro', linestyle='-', linewidth=0.5)
+ax2.plot(Bubble[:, 1]*1e6, Bubble[:, 4], linestyle='solid', linewidth=0.75,color='navy')
+
+ax3.set_yscale('log')
+ax3.set(xlabel=r'$t$ [$\mu$s]',ylabel=r'$p_\mathrm{G}(t)$ [Pa]')
+ax3.set_xlim(xmin=6,xmax=11)
+ax3.grid(color='gainsboro', linestyle='-', linewidth=0.5)
+ax3.plot(Bubble[:, 1]*1e6, (Bubble[:, 5]), linestyle='solid', linewidth=0.75,color='navy')
+
+ax4.set_xscale('log')
+ax4.set_yscale('log')
+ax4.set(xlabel=r'$r$ [m]',ylabel=r'$\Delta p (r,t)$ [MPa]')
+ax4.set_xlim(xmin=1e-7,xmax=1e-3)
+ax4.set_xticks([1e-7,1e-6,1e-5,1e-4,1e-3])
+ax4.set_ylim(ymin=0.1,ymax=1e4)
+ax4.set_yticks([0.1,1,1e1,1e2,1e3,1e4])
+ax4.grid(color='gainsboro', linestyle='-', linewidth=0.5)
+ax4.plot([3e-6,8e-4], [40,0.15], color='gray', linestyle='dashed', lw=0.7)
+ax4.annotate('$-1$', fontsize = 8, color='gray', xy=(0.35,0.3), xycoords='axes fraction')
+ax4.plot(pLmax[:, 1], (pLmax[:, 3]-pLmax[:, 6])*1e-6,  linestyle='solid', linewidth=0.75,color='navy')
+
+ax5.set_xscale('log')
+ax5.set_yscale('log')
[…]
