### 2023-06-02 7be49bb
Added a script to run and plot all examples
