+++ README.md
@@ -11,7 +11,10 @@
     <img src="https://img.shields.io/badge/DOI-10.5281/zenodo.7249297-blue" alt="Latest version">
   </a>
   <a href="https://doi.org/10.1063/5.0131930">
-    <img src="https://img.shields.io/badge/Paper-10.1063/5.0131930-blue" alt="Paper">
+    <img src="https://img.shields.io/badge/PoF-10.1063/5.0131930-blue" alt="Paper">
+  </a>
+  <a href="https://joss.theoj.org/papers/27166cd5496c33d62e74132712efec8a">
+    <img src="https://joss.theoj.org/papers/27166cd5496c33d62e74132712efec8a/status.svg">
   </a>
 </p>
 
@@ -33,28 +36,35 @@ Key features of APECSS are:
 APECSS is under the copyright of its developers and made available as open-source software under the terms of the [Mozilla Public License Version 2.0](LICENSE.txt).
 
 ## Quick Start Guide
+
+### **Installation**
 Getting started with APECSS is easy. After downloading APECSS in the directory ````<path to APECSS>````, define the following environment variables:
 - ````APECSS_DIR```` to the directory in which APECSS is located. Using bash, for instance, simply execute the command ````export APECSS_DIR=<path to APECSS>```` or, even better, add this command to your bash profile.
 - ````USRLIB_DIR```` to the directory in which libm.a or libm.dylib (the standard _math_ library) is located. This may, for instance, be ````/usr/lib64/```` on Linux systems or ````/usr/lib/```` on MacOS systems.
 
 Now, navigate into the folder ````$APECSS_DIR/lib```` and execute ````./compile_lib.sh````. This shell script will compile the APECSS library using _cmake_ with the ````CMakeLists.txt```` file provided in this folder. By default, APECSS is compiled with double precision and in _Release_ mode, meaning all optimization flags are enabled. That's it, you've successfully compiled APECSS!
 
+### **Examples**
 There are several ways in which you can use the APECSS library. You can either incorporate selected features of APECSS into your own software or you can program an interface to use APECSS as a standalone program. Some representative examples are given in the [````$APECSS_DIR/examples````](/examples/) directory. Each directory contains the following:
 - A ````README.md```` file explaining the purpose and specificities of this/these example(s).
 - A ````src```` folder with a file called ````*_apecss.c```` that acts as the standalone interface to the APECSS library. This file contains the ````main()```` function and any additional functionality required to simulate a specific scenario.
 - A ````build```` folder containing the ````CMakeLists.txt```` file and a shell script ````compile.sh```` with which this example can be compiled using the command ````./compile.sh````.
 - One or several ````*.apecss```` files in which the options for a specific case are defined.
+- One or several ````plot_*.py```` Python scripts to plot the results of the example. These Python scripts require ````numpy```` and ````matplotlib````, and plot the results into a pdf-file. 
+- Some examples have a ````reference-data```` folder, which contains results of previous studies to quickly validate the overall correct working of APECSS and which are plotted alongside the results produced by APECSS using the Python script(s).
 
-If you'd like, for instance, test the installation of APECSS with an ultrasound-driven bubble, navigate to [````$APECSS_DIR/examples/ultrasound````](/examples/ultrasound/). There you can find different test cases, described in more detail in the accompanying [````README.md````](/examples/ultrasound/README.md). Execute the ````compile.sh```` script in the [````build````](/examples/ultrasound/build/) directory. You can now choose which test case to run, by using the corresponding execution command suggested in the [````README.md````](/examples/ultrasound/README.md). 
[…]
