+++ documentation/chapters/using.tex
@@ -51,11 +51,17 @@ Some representative examples are available in the {\tt \$APECSS\_DIR/examples} f
   \item A {\tt src/} folder with a file called {\tt *\_apecss.c} that acts as the standalone interface to the APECSS library. This file contains the {\tt main()} function and any additional functionality required to simulate a specific scenario.
   \item A {\tt build/} folder containing the {\tt CMakeLists.txt} file and a shell script {\tt compile.sh} with which this example can be compiled using the command {\tt ./compile.sh}.
   \item One or several {\tt *.apecss} file(s) in which the options for a specific case are defined.
+  \item One or several {\tt plot\_*.py} Python scripts to plot the results of the example. These Python scripts require {\tt numpy} and {\tt matplotlib}, and plot the results into a pdf-file. 
+  \item Some examples have a {\tt reference-data} folder, which contains results of previous studies to quickly validate the overall correct working of APECSS and which are plotted alongside the results produced by APECSS using the Python script(s).
 \end{itemize}
 
-In the examples provided in the {\tt \$APECSS\_DIR/examples} folder, the name of the  {\tt *.apecss} file is passed as an argument with the call to run APECSS, e.g.~executing {\tt ./<APECSS-example> -options run.apecss} to use an options file named {\tt run.apecss}.
+In the examples provided in the {\tt \$APECSS\_DIR/examples} folder, the name of the  {\tt *.apecss} file is passed as an argument with the call to run APECSS, e.g.~executing {\tt ./<APECSS-example> -options run.apecss} to use an options file named {\tt run.apecss}. Note that, for most test cases, the settings and options given by the execution command and the corresponding {\tt *.apecss} file reproduce a case previously presented in the literature, as indicated in the {\tt README.md} files, which also serves to validate the implementation of APECSS.
 
-Detailed information about each example, how to run it and what the results might be compared to can be found in the accompanying {\tt README.md} file.
+You can run \textbf{all} examples by executing {\tt ./run\_all.sh} in the {\tt examples} folder. This shell script compiles, runs and plots the results of all test cases. This is a great way to test the installation of APECSS and see what sort of results can be output and how these results may be visualized. Note that {\tt run\_all.sh} assumes that \textit{Python} can be executed using the command {\tt python3}; please adjust this if necessary.
+
+Alternatively, if you're, for instance, specifically interested in an ultrasound-driven bubble, navigate to {\tt \$APECSS\_DIR/examples/ultrasound}. There you can find different test cases, described in more detail in the accompanying {\tt README.md}. Execute the {\tt compile.sh} script in the {\tt build} directory. You can now choose which test case to run, by using the corresponding execution command suggested in the {\tt README.md}. 
+
+More information about each example, how to run it and what the results might be compared to can be found in the accompanying {\tt README.md} file.
 
 \section{Programming in and with APECSS}
 
@@ -297,13 +303,15 @@ Various simulation results associated with the bubble dynamics and the acoustic 
 
[…]
