### 2023-06-02 #5435 @fabiandenner
I updated the examples and documentation as suggested by @svchb. The in-code documentation of the source files has been improved and every example now comes with a Python script to plot the results, including reference results from the literature where applicable, which also allows a quick validation of the installation. In addition, the example folder now also contains a shell script, ````run_all.sh````, which compiles, runs and plots the results of all examples. I updated the README.md and the pdf-documentation accordingly. Sample results can now also be found in the pdf-documentation.
