### 2023-06-13 polycfd/apecss#14 @jmansour
JOSS / Explicitly list requirements for library
Please add an explicit "Requirements" section to your `README.md` listing dependencies, including the necessary compiler.

Please also add optional requirements which may be necessary for plotting results (Python/Numpy/Matplotlib) and testing.
