### 2023-06-13 polycfd/apecss#15 @jmansour
JOSS / quantitative test
Please add a simple quantitative test, possibly to be executed via the `run_all.sh` script.

This could be something as simple as a `numpy.allclose()` call within one of the plotting routines to ensure produced results are within some tolerance of expected results.
