### 2023-06-13 #5435 @jmansour
Hi all

Apologies for my response coming in at the taller end of the review time window. I was hoping to get it done sooner, but alas.

Firstly I'd like to commend the authors on this excellent contribution to scientific OSS canon. The library looks well written, using modern C paradigms and modularity, and this should bode well for the project's sustainability and future growth. The implemented functionality appears to be appropriately expressed via the API, and this is reflected in the provided examples which are clean, concise and easy to digest.

The library is extremely lightweight with minimal dependencies. This is a very advantageous for both end users and also third parties wanting to use your library to add cavitation dynamics to existing applications. Following the provided instructions, I found it trivial to build the library on Ubuntu and OSX. While the front Github page does state the dependencies within the text, I would like to see an explicit list provided under a "Requirements" section, and you should also list the _optional_ requirements for plotting results (Python/numpy/matplotlib).
https://github.com/polycfd/apecss/issues/14

The library appears to fill an important gap in scientific OSS. As discussed earlier in this thread, there doesn't appear to be any alternative open source offerings which explicitly model bubble dynamics and cavitation, which is indeed surprising given the maturity of the governing dynamics and breadth of application (as outlined within the author's JOSS submission). Fluent does appear to provide some cavitation functionality, but given it's a commercial package I don't think it needs to be considered here as an alternative.

Most of the examples provided within the repository reproduce results available in the literature:
 * examples/binaryinteraction - Jiang et al 2017 (qualitative)
 * examples/gastemperature - Stricker et al 2011 (qualitative)
 * examples/laserinducedcavitation - Liang et 2022 (provided data)
 * examples/rayleighcollapse - Denner & Schenke 2023 (provided data)
 * examples/ultrasound_lipidcoated_simple - Marmottant et al 2005 (qualitative)
 * examples/ultrasound_sonolum - Holzfuss 2010 (provided data)
 * examples/ultrasound_kelvinvoigt - Yang & Church 2005 (qualitative)
 * examples/ultrasound_zener - Zilonova et al 2018 (qualitative)
 * examples/ultrasound_oldroydb - Jimenez-Fernandez & Crespo 2005 (qualitative)
I was able to reproduce all the above results locally, and I am satisfied that this is a sufficient demonstration of the correctness of the implementation. Note that I have trusted that the third party reference data provided within the repository is accurate.

The JOSS submission includes the results produced by the `ultrasound_sonolum` example, and this was able to be reproduced, although I also note that these are not original results so reproduction is perhaps not strictly required here.

I was able to use the `run_all.sh` script to automatically run all the provided examples, although as far as I can tell these do not perform any quantitative tests. I would like to see a simple quantitative test added as this is necessary for the verification of the software's functionality.
https://github.com/polycfd/apecss/issues/15

With regards to the JOSS software paper itself, I am satisfied that it meets all the listed requirements.
