### 2023-06-13 #5435 @jmansour
Some further thoughts/suggestions I have which you might consider for future work, although shouldn't block the JOSS acceptance:

* Generating Python bindings for your API would hugely increase accessibility for a large section of the modelling community, and generally wouldn't incur any appreciable performance overhead. I think this would be a relatively straightforward undertaking via tools such as Cython or SWIG given your library's design and minimal requirements.
* Consider dropping CMake, as it's a relatively heavy build system and overkill for your use case. Make would be entirely sufficient, and is generally available as default, but you could even have your compile scripts execute the required compilation sequence directly.
* Confidence could certainly be improved by providing datasets for the models where only qualitative comparison is currently possible. Automated tests for those too! 🙂
* You might considering using an automated API documentation tool such as Doxygen, as manually ensuring your documentation stays in sync with your code-base is difficult and error prone.
* I generally prefer C-libraries to by default generate _debug_ builds, as this is generally what you will require during model construction, with release/optimised builds used for production runs. Though I note that you have provided instructions for debug builds within your documentation.
* You should add shebangs to your scripts to ensure they are executed by the appropriate interpreter.
