+++ documentation/chapters/using.tex
@@ -1,5 +1,11 @@
 \chapter{Using APECSS}
 
+\section{Requirements}
+
+APECSS has been developed and tested on Unix systems, using Linux and MacOS operating systems. The only mandatory requirement to compile and run APECSS is a standard C compiler, such as {\tt gcc}.
+
+Optionally, {\tt cmake} (version 3.12 or higher) is required if you would like to make use of the provided compilation and test scripts. To use the provided Python scripts for visualizing the output of APECSS, {\tt Python} (version 3), {\tt numpy} and {\tt matplotlib} are required.
+
 \section{Installation}
 \label{sec:installation}
 
