### 2023-06-13 polycfd/apecss#14 @polycfd
Nice idea! I've added a requirements section to the main README.md and the documentation (see Section 2.1).
