### 2023-06-13 #5435 @fabiandenner
@jmansour Thank you very much for the detailed review and the excellent suggestions. No need to apologize for your review "coming in at the taller end of the review time window" - I prefer a constructive and diligent review over a rushed response. I'm glad to hear that you have had a good time exploring our software library.

Your suggestions are very much appreciated. I have just committed a small update including a _Requirements_ section in the main README.md and in the pdf documentation (see Section 2.1). As we have been using this library actively for our research, we have taken the utmost care in implementing the various model correctly and given the breadth of previously published results APECSS reproduces accurately, we are confident that we have succeeded in that quest. Nevertheless, the automated quantitative tests you suggested are a very good idea. I'll try to find some time in the coming days to explore how we can best incorporate such quantitative tests, for instance, as you suggested, using ````numpy.allclose()```` in the ````run_all.sh```` script.

Regarding your further thoughts, we have already been thinking about introducing bindings for Python and Fortran, but haven't had the opportunity to pursue this further. I absolutely agree with you that this would increase the accessibility significantly. We considered CMake for convenience and especially during the development this simplifies things. However, we will keep adding alternatives to CMake (i.e. providing simple, ready to go Makefiles) in mind as a future addition. Automated tests are already in development, as mentioned above. Regarding the shebangs, thanks for pointing this out to us. We've simply forgotten to add them because the scripts run with both ````bash```` and ````zsh```` shells. We'll add them shortly.

Thanks again for the helpful suggestions!
