### 2023-06-14 0f0abc5
Adding automated results test for Rayleigh collapse with emissions
