+++ .github/workflows/test_run.yml
@@ -15,6 +15,13 @@ jobs:
     runs-on: macOS-latest
     steps:
     - uses: actions/checkout@v3
+    - uses: actions/setup-python@v4
+      with:
+        python-version: 3.11
+    - name: Configure Python
+      run: |
+        python -m pip install numpy
+        python -m pip install matplotlib
 
     - name: Configure lib
       run: cmake -S ${{github.workspace}}/lib -B ${{github.workspace}}/lib -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}}
@@ -48,8 +55,10 @@ jobs:
       run: cmake --build ${{github.workspace}}/examples/rayleighcollapse/build --config ${{env.BUILD_TYPE}}
     - name: Run Rayleigh collapse (simple)
       run: ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/simple.apecss -tend 1
-    - name: Run Rayleigh collapse (with emissions)
-      run: ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/emissions.apecss -tend 0.2
+    - name: Run Rayleigh collapse (with emissions) and check results
+      run: |
+        ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/emissions.apecss -tend 0.2 -resultsdir ${{github.workspace}}/examples/rayleighcollapse/
+        python plot_result_emissions.py ${{github.workspace}}/examples/rayleighcollapse/
 
     - name: Configure ultrasound
       run: cmake -S ${{github.workspace}}/examples/ultrasound/build -B ${{github.workspace}}/examples/ultrasound/build -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}}
@@ -72,6 +81,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
+    - uses: actions/setup-python@v4
+      with:
+        python-version: 3.11
+    - name: Configure Python
+      run: python -m pip install --upgrade pip requests
 
     - name: Configure lib
       run: cmake -S ${{github.workspace}}/lib -B ${{github.workspace}}/lib -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}}
@@ -105,8 +119,10 @@ jobs:
       run: cmake --build ${{github.workspace}}/examples/rayleighcollapse/build --config ${{env.BUILD_TYPE}}
     - name: Run Rayleigh collapse (simple)
       run: ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/simple.apecss -tend 1
-    - name: Run Rayleigh collapse (with emissions)
-      run: ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/emissions.apecss -tend 0.2
+    - name: Run Rayleigh collapse (with emissions) and check results
+      run: |
+        ${{github.workspace}}/examples/rayleighcollapse/build/rayleigh_apecss -options ${{github.workspace}}/examples/rayleighcollapse/emissions.apecss -tend 0.2 -resultsdir ${{github.workspace}}/examples/rayleighcollapse/
+        python plot_result_emissions.py ${{github.workspace}}/examples/rayleighcollapse/
 
     - name: Configure ultrasound
       run: cmake -S ${{github.workspace}}/examples/ultrasound/build -B ${{github.workspace}}/examples/ultrasound/build -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}}
