### 2023-06-14 2675569
Bug fix in GithubActions script
