### 2023-06-14 1d24178
Bug fix in the plot script for the Github Actions runner
