### 2023-06-14 e3d4a5c
Test if the build action fails as it should with bad results
