+++ documentation/chapters/using.tex
@@ -63,7 +63,7 @@ Some representative examples are available in the {\tt \$APECSS\_DIR/examples} f
 
 In the examples provided in the {\tt \$APECSS\_DIR/examples} folder, the name of the  {\tt *.apecss} file is passed as an argument with the call to run APECSS, e.g.~executing {\tt ./<APECSS-example> -options run.apecss} to use an options file named {\tt run.apecss}. Note that, for most test cases, the settings and options given by the execution command and the corresponding {\tt *.apecss} file reproduce a case previously presented in the literature, as indicated in the {\tt README.md} files, which also serves to validate the implementation of APECSS.
 
-You can run \textbf{all} examples by executing {\tt ./run\_all.sh} in the {\tt examples} folder. This shell script compiles, runs and plots the results of all test cases. This is a great way to test the installation of APECSS and see what sort of results can be output and how these results may be visualized. Note that {\tt run\_all.sh} assumes that \textit{Python} can be executed using the command {\tt python3}; please adjust this if necessary.
+You can run \textbf{all} examples by executing {\tt ./run\_all.sh} in the {\tt examples} folder. This shell script compiles, runs and plots the results of all test cases. This is a great way of testing the installation of APECSS and see what sort of results can be output and how these results may be visualized. The script also compares the results computed for a respresentative example (the Rayleigh collapse with emissions) with reference results of APECSS, checking whether the installation produces correct results. Note that {\tt run\_all.sh} assumes that \textit{Python} can be executed using the command {\tt python3}; please adjust this if necessary.
 
 Alternatively, if you're, for instance, specifically interested in an ultrasound-driven bubble, navigate to {\tt \$APECSS\_DIR/examples/ultrasound}. There you can find different test cases, described in more detail in the accompanying {\tt README.md}. Execute the {\tt compile.sh} script in the {\tt build} directory. You can now choose which test case to run, by using the corresponding execution command suggested in the {\tt README.md}. 
 
