+++ README.md
@@ -59,7 +59,7 @@ There are several ways in which you can use the APECSS library. You can either i
 - One or several ````plot_*.py```` Python scripts to plot the results of the example. These Python scripts require ````numpy```` and ````matplotlib````, and plot the results into a pdf-file. 
 - Some examples have a ````reference-data```` folder, which contains results of previous studies to quickly validate the overall correct working of APECSS and which are plotted alongside the results produced by APECSS using the Python script(s).
 
-You can run **all** examples by executing ````./run_all.sh```` in the [````examples````](/examples/) folder. This shell script compiles, runs and plots the results of all test cases. This is a great way to test the installation of APECSS and see what sort of results can be output and how these results may be visualized. Note that ````run_all.sh```` assumes that *Python* can be executed using the command ````python3````; please adjust this if necessary.
+You can run **all** examples by executing ````./run_all.sh```` in the [````examples````](/examples/) folder. This shell script compiles, runs and plots the results of all test cases. This is a great way of testing the installation of APECSS and see what sort of results can be output and how these results may be visualized. The script also compares the results computed for a respresentative example (the [Rayleigh collapse with emissions](/examples/rayleighcollapse/)) with reference results of APECSS, checking whether the installation produces correct results. Note that ````run_all.sh```` assumes that *Python* can be executed using the command ````python3````; please adjust this if necessary.
 
 Alternatively, if you're, for instance, specifically interested in an ultrasound-driven bubble, navigate to [````$APECSS_DIR/examples/ultrasound````](/examples/ultrasound/). There you can find different test cases, described in more detail in the accompanying [````README.md````](/examples/ultrasound/README.md). Execute the ````compile.sh```` script in the [````build````](/examples/ultrasound/build/) directory. You can now choose which test case to run, by using the corresponding execution command suggested in the [````README.md````](/examples/ultrasound/README.md). Also, the settings and options given by the execution command and the corresponding ````*.apecss```` file typically reproduce a case previously presented in the literature, as indicated in the ````README.md```` file.
 
