### 2023-06-14 polycfd/apecss#15 @fabiandenner
Thanks for this suggestion. A quantitative test is now implemented.
