### 2023-06-15 #5435 @jmansour
Nice work @fabiandenner! Sometimes it can indeed be challenging to ensure tests remain robust across a number of platforms. The options you mention are a good idea, although sometimes simply increasing the error tolerance is also fine (if perhaps less satisfying). Also, I came across the following paper at some point which provides an analytic solution for RP governed bubble growth. Maybe worth a look, as an analytic solution would be an excellent addition to your library's validation portfolio:
https://amsemodelling.com/publications/advances_in_modelling/General_Mathematics/552/55.02_07.pdf

That covers everything from my end. Happy for the publication to proceed. @kyleniemeyer

Best with the project @fabiandenner.
