### 2023-06-15 #5435 @fabiandenner
@svchb and @jmansour
Thank you both, also on behalf of my co-author @SoerenSchenke, for the helpful review and the constructive suggestions. This was certainly a worthwhile experience since, aside from improving the quality of the software project, we also learned about how to make open-source projects more accessible.

@jmansour Thanks for the suggestion of the analytical solution, we will have a look into it. Comparing against an analytic solution was the main motivation for the _Rayleigh collapse simple_ example. (i) All dissipation mechanisms are neglected, meaning the bubble should oscillate with constant amplitude, and (ii) the collapse time may be compared with the analytical collapse time proposed by Rayleigh. APECSS reproduces both accurately.
