### 2023-06-15 834c62a
Corrected DOI of Hickling & Plesset (1963) reference in paper.bib
