+++ paper/paper.bib
@@ -504,7 +504,7 @@
   number = {85-24},
   address = {{Pasadena, California, USA}},
   institution = {{California Institute of Technology}},
-  url = {https://resolver.caltech.edu/CaltechAUTHORS:HICpof64}
+  url = {https://resolver.caltech.edu/CaltechAUTHORS:DivEngAppSciRpt85-24}
 }
 
 @article{Holzfuss2010,
