### 2023-06-15 #5435 @fabiandenner
@kyleniemeyer I noticed two small mistakes while proof reading the paper just now.

1. The DOI of the referenced report of Hickling & Plesset (1963) was incorrect (pointed to a different publication of the same authors.
2. The y-axis of Fig. 1(f) was incorrectly labelled as $\Delta p(r,t)$, whereas it should be labelled as $\Delta p(t)$, see caption.

I corrected both.
