### 2023-06-15 #5435 @fabiandenner
@kyleniemeyer One of the missing DOIs, for the book _Underwater Explosions_ is spot on, well done editorialbot. I've added it to the bib file. Regarding the bot-suggested DOI, for the report  _The Collapse of a Spherical Cavity in a Compressible Liquid_, it does not seem to work. However, the Cal Tech address in the reference is supposedly a persistent identifier provided by Cal Tech.
