### 2023-06-15 f0f827f
Updated the references to the JOSS paper
