+++ documentation/chapters/using.tex
@@ -363,11 +363,11 @@ The {\tt Build test} checks whether the APECSS library compiles correctly on Lin
 
 \section{Citing APECSS}
 
-If you use APECSS for your scientific work, please consider citing the paper introducing the theoretical foundation of APECSS,\vspace{-0.5em}
+If you use APECSS for your scientific work, please consider citing the paper introducing the features and capabilities of APECSS,\vspace{-0.5em}
 \begin{itemize}[leftmargin=*]
-  \item[] F.~Denner and S.~Schenke, Modeling acoustic emissions and shock formation of cavitation bubbles. \textit{Physics of Fluids} 35 (2023), 012114. \href{https://doi.org/10.1063/5.0131930}{\texttt{https://doi.org/10.1063/5.0131930}} 
+  \item[] F.~Denner and S.~Schenke, APECSS: A software library for cavitation bubble dynamics and acoustic emissions. \textit{Journal of Open Source Software} 8 (2023), 5435. \\ \href{https://doi.org/10.21105/joss.05435}{\texttt{https://doi.org/10.21105/joss.05435}} 
 \end{itemize}\vspace{-0.5em}
-as well as the version of APECSS you've used for your work, e.g.\vspace{-0.5em}
+as well as the version of APECSS you've used for your work, for instance\vspace{-0.5em}
 \begin{itemize}[leftmargin=*]
   \item[] F.~Denner and S.~Schenke, APECSS (v1.2), \textit{Zenodo} (2022). \\ \href{https://doi.org/10.5281/zenodo.7465050}{\texttt{https://doi.org/10.5281/zenodo.7465050}}
 \end{itemize}\vspace{-0.5em}
