+++ CITATION.cff
@@ -1,0 +1,34 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Denner
+  given-names: Fabian
+  orcid: "https://orcid.org/0000-0001-5812-061X"
+- family-names: Schenke
+  given-names: Sören
+  orcid: "https://orcid.org/0000-0001-8765-3722"
+doi: 10.5281/zenodo.8043400
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Denner
+    given-names: Fabian
+    orcid: "https://orcid.org/0000-0001-5812-061X"
+  - family-names: Schenke
+    given-names: Sören
+    orcid: "https://orcid.org/0000-0001-8765-3722"
+  date-published: 2023-06-15
+  doi: 10.21105/joss.05435
+  issn: 2475-9066
+  issue: 86
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5435
+  title: "APECSS: A software library for cavitation bubble dynamics and
+    acoustic emissions"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05435"
+  volume: 8
+title: "APECSS: A software library for cavitation bubble dynamics and
+  acoustic emissions"
