### 2023-06-15 c5e2d65
Fixed a typo in the README.md
