### 2023-06-15 #5435 @fabiandenner
@kyleniemeyer Thanks for wrapping this up and thanks to @svchb and @jmansour for taking the time to review our paper and code.

I hope other journals take notice of the excellent workflow JOSS is running, very good work!
